Split out user types; add oauth user auth
This moves the user code into the 'types' mod to be shared with the
upcoming api server.

This also adds a from request implementation to auth a user via
authorization token.
